+++ GitAutomation.Core/Orchestration/IOrchestrationActions.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using GitAutomation.Processes;
+
+namespace GitAutomation.Orchestration
+{
+    public interface IOrchestrationActions
+    {
+        IObservable<OutputMessage> CheckDownstreamMerges(string downstreamBranch);
+        IObservable<OutputMessage> ConsolidateServiceLine(string releaseCandidateBranch, string serviceLineBranch, string tagName);
+    }
+}
